requirements.txt
beautifulsoup4
black
tiktoken
openai
unstructured
nltk
fastapi
jinja2
uvicorn
streamlit
streamlit-chat
tqdm
isort
langchain
langchainhub
langchain-community
langchain-google-genai
langsmith
pinecone-client
pillow
google-generativeai
firecrawl-py
python-dotenv
